Use Logos to generate a lexer
It seems to do a good job though I don't know for sure if it'll be able
to deal with everything we want to throw at it but it seems ok.

We also add 'Main.elm' from the elm-syntax-sscce project:

  https://github.com/pdamoc/elm-syntax-sscce/

To give us a complete example of the syntax to parse.

The lexer currently doesn't deal with triple-quote strings but that is
possibly not too hard. We'll see.