Add basic evalulator
Really it just finds the 'main' function and prints the expression
associated with it. But it is a start and it allows us to get some kind
of 'value' out of our Basic.elm script.

We also change the dump.rs file so that it returns an Error from 'main'
so that we can more easily early exit and pass values around, etc.