Broken support for if-expression
We also remove the lifetime from the parser::Error so that we can pass
it around more easily.

We also change the parse function to receive the iterable directly
rather than the source tokens as that brings it inline with the other
functions in the file.

The if-expression test is failing with an unexpected token and I'm
hoping to take the time to get more details out of the error but that
involves dealing with source spans so I'm committing this broken state
now and hopefully we'll get the spans working first and then debug this
issue with that.