Fix if-statement parsing
We start tracking the token ranges and use the range information in our
errors with the codespan-reporting library to get clearer test errors.

We then rethink our 'Context'/'indentation tracking' approach as we seem
to need to keep track of the 'current' indentation in a manner that is
hard to express with the Contexts & child Contexts. I imagine there is a
less verbose way to do what we're doing here but it seems to work for
some if-statements and nested if-statements so that's cool.