Support args for main function
Early days but it seemed like a bit of fun to add support for passing
something into the program as it gives us something to manipulate and
play with. Or rather to just return directly as there isn't much to do
with a string list.

We do this by implementing some kind of support for function calls and
then calling the 'main' function with any provided args if args are
provided.

If args aren't provided and the main declaration doesn't have args then
we treat it as a 'binding' instead and just evaluate the body expression
directly as before.