Further parsing hiearchy changes
We don't think that if-statements can be only side of a binary without
parenthesis so it doesn't make sense to look for them at the same level
as a binary statement which means that we can move if-statement and
binary statement parsing to the same entry level.

I think this is the approach in Crafting Interpreters.