Restructure indentation values
So that the use a single field for the status and another for the count.
This means we don't always have to open up the enum entries to get to
the count and whilst they both want to store the 'count' there is no
benefit in having it integrated into both.