Reorganise modules
We rename 'evaluate' to 'evaluater' for consistency with the other
modules.

We also change the 'dump' entry point to use 'erm' as a crate so that we
only need to have the 'mod' declarations in lib.rs. I'm still getting
used to all of this.

We also move 'env' out to its own file so that we can use it from other
modules more easily. Mostly for the type checker module though I don't
it'll be share long term.