Add first attempt at a type checker
I'm quite confused by how I'm meant to go about this. My current
assumption is that we must decide on an expected type for the main
function and then try to 'unify' the terms for that expected type with
the terms for the body of the main function.

This is based partly on poking at the Elm compiler and seeing it
complain about needing Main to be of a particular type. I guess it can
be of anything really but it feels like we need to have something for it
in order to do the unification process. Or perhaps we could unify it
with a generic type and then see what the generic type ends up being.

We're not handling the case where the main function takes arguments. And
most of the tests currently fail.

In order to get some of the basic tests to pass we have to introduce a
'stringFromInt' function which we treat as a built in for the moment. As
it is dealing with primitive types it is always going to have to be in
the run time but we don't have a solid built-in function concept yet so
we've cobbled together something which seems adequate for the moment but
not that grand.

It also introduces an import cycle between env & evaluater that I don't
love but at least it compiles for the moment.