Lookup term for built-in function
We want to be able to look up the signature terms for the built-in
functions that we have instead of hardcoding it into the checker code.

This achieves and then also moves a lot of stuff around to try to reduce
cyclic imports as they make me unhappy. Unfortunately the built in
function uses stuff from parser, checker and evalulator so it becomes
necessary to break them all out a bit in order to avoid the cyclic
imports.

I worry that having them all in different files in not good for code
ergonomics but I'm not sure what else to do right now.