Start to include 'Basics.elm' file
There is probably a better place for this long term but for the moment
it is interesting to start thinking about a 'prelude' the default stuff
that is available for any Elm code and whilst we could declare all of
that as built-in functions and data, it is interesting to try to express
it as a Basics.elm file with code that we loads the functions, types and
operators that we're interested in.

Unfortunately this breaks everything as we don't recognise 'infix'
statements but we can work on that.