Change module & scope approach
Instead of passin both the module & scopes around we create a scope from
the module and then just pass scopes around. This allows us to unify
things a little more and be less dependent on the module as some kind of
source of truth through execution.

It also makes it easier to introduce the 'Basic.elm' contents or
'prelude' (I hope) by having a scope in the stack before even the
current module. We'll see how that goes.