Switch to Rc's for all expressions
So that it is easier to reference them without worrying about the
lifetimes of the references.

The only lifetime we're now mostly interesting in caring about is the
lifetime of the source code itself as we still reference that in a lot
of places.

This also includes initial work on checking binary expressions and
operators as that is what I was working on when I lost all sense of
direction with the reference approach and, now that things are compiling
again, I just want to commit something before continuing with the
operator work.