A messy start to type-checking
Comments and debugging print statements. A number of the tests are
passing though. I haven't finalised 'main with args' yet and what we
have is full of bugs including the failing test because we can't define
our own function called 'add' because we'll find the 'add' in Basics
first, somehow, so we need to fix that too.

My plan is to clean it all up and comment it properly before it goes
into main.