Store operator function with operator
So that it is easier to retrieve the implementation of the operator when
checking & evaluating it. Previously when looking up the 'add' function
associated with the '+' operator it would find the 'add' function in the
scope of the current module and not the one in the module that defined
'+'.

This isn't the best fix as it doesn't generalise to other functions that
call functions but it seems reasonable to do in the case of operators.