Fix argument & scope handling
We evaluate the arguments to values before adding them to the scope that
is being provided for the function body evaluation. This means that we
avoid an infinite loop when passing in an argument that is itself an
argument which mean that it was a VarName that evaluated to a VarName.

We partly achieve this by changing our approach to scopes by realising
that we need to be less simplistic than a simple stack of scopes. For
example, when calling a function we need to be able to pass in a fresh
environment that includes any bindings for the arguments and also the
scope for the module itself but not for any intermediate scopes around
the function call itself. A function shouldn't know about the local
variables around the call site.

Consequently we change from 'Scopes' to 'Environment' and split out the
'module' and 'local' levels. I'm sure there will be lots more to adjust
with time but for the moment all the tests pass.