Switch from stringJoin to String.join
To match the Elm standard library, now that we can.

We also realise that there is no real different between eval and
eval_with_args as test functions and so one can be implemented with the
other. Not sure why it took me so long to notice.

eval_with_args was a bit behind so I've updated that to match 'eval' and
include the prelude and everything.