Check & evaluate bindings in their local environment
We had an issue where we were importing two layers deep and the second
import was failing due to trying to evaluate it in the environment of
the first file.

The system isn't perfect yet, I doubt, but we attempt to return the
relevant environment for any binding that we look up so that if we're
checking or evaluating it then we can do so in the correct environment
instead of in the initial environment that we've been passing down the
stack.

This cleans up the binding types where we originally had BuiltIn as
part of the main bindings just so we could return it from the
get_binding function even though we never create builtin binding when
parsing the source code.

Now we've more appropriately separated it out so it is just in the type
of the get_binding function and the other bindings stay together.

We also change the regex for LowerName to include underscores as our
example includes them and highlighted that our lexing was failing to
handle them.