Improve imports, scopes and environment
We haven't been respecting the exposing part of the import statement in
various ways. We introduce 'ModuleImport' as a concept to try to wrap up
a module struct with some extra data that describes with the specific
import exposes.

This generally establishes a clearer model for modules, imports and
their various scopes. Probably not perfect but better than before.