Support empty list syntax
And some kind of type checking around it.

We change the definition of String.join to include a string that does
the joining. Not that is is useful for the empty list situation but it
matches the actual definition of 'String.join' in elm/core instead of
hard coding the joiner to be " ".