Require args for main
We probably want to change this in future to require a type statement
for the main function and then to build from that or maybe not even as
programs basically only launch with a set of string arguments.

Either way it is silly to expect the definition of main to vary
depending on how many run-time arguments the user is providing which is
what was happening before. If they provided no args on the command line
then we'd expect a variable called main instead of a function called
main but that is nuts. Instead if they provide no args then we pass
through an empty list (or that is the expectation.)