Read from stdin if no file
Part of this is slightly wrong. I wanted to be able to do #! files with
erm and I thought that #! meant that the code was sent into standard
input but it seems that that is not the case. Instead the interpreter is
given the path to the file that #! at the top of it. This means we have
to strip out the #! line before passing it into the evaluation mechanism
as the #! line is not valid Elm.

The stdin also seems to work though which is nice. Respects ctrl-d as
end of content.