Support partially applied functions & snapshot testing
We introduce a test with the syntax:

  main args =
      String.fromInt ((add 2) 3)

Where the parens cause the 'add 2' to be evaluated before the 3 comes
along to complete the arguments for add. We weren't supporting it
before. We introduce a 'PartiallyAppliedFunc' value into the evaluation
so that we can store a function and the arguments that it has seen so
far.

That accounts for most of the changes to the core code, the rest is due
to introducing snapshot changes where we both use the inst library for
snapshotting the test output to reduce maintenance and also switch to
having a string output for the test eval function where we use the
error::to_user_output function that we started to prepare for the run.rs
execution and we attempt to make sure that it shows string output
as expected (which is the only valid success output at the moment) along
with some more reasonably formatted errors including use of the
codespan_reporting library to pretty print parsing issues.

We also:

- Remove the non-args main binding option from the evaluation.

- Avoid returning the 'dyn Func' from the get_bindings helper and
  instead just return the built-in string identifier for the function
  which is resolve at a later point. It seems less messy than dealing
  with passing 'dyn Func' around more than we want (and we would be
  passing it around more than we did before due to the mechanics for
  handling partial application.)

- Clear out the helpers that we're in 'lib.rs'. They were sort of useful
  but not useful or familiar enough to me.

- Add more trace logging through the code.

- Try to remove line numbers from error enums in favour of using
  log::error! statements that have been updated to include the file and
  line number in the output so as long as we log before each error then
  we're ok.