Use 'Qualified' phrasing instead of 'Path'
Still just in the lexer for the moment.

Also include comments and allow the QualifiedLowerName to include to
include more lower case segments. So:

  MyModule.myRecord.myField

Instead of just:

  MyModule.myValue

That said, we don't have it quite right here as the LowerName should
allow that too as the QualifiedLowerName requires a capitalised module
name to be present.