Differentiate names in the ast
We have UpperName and QualifiedUpperName and LowerName and
QualifiedLowerName. It might not pay off exactly but I think it ok.

We have to move the Lexer back to LowerName and LowerPath and
UpperName and UpperPath as we can't have the otherwise required overlap
between QualifiedUpperName and UpperName, etc, at the lexer level as
the logos library doesn't like it.

So we have one way to distinguish in the lexer and another in the ast.
Hopefully it'll be useful still. Changing the ast allows us to say that,
for example, a statement function or binding name shouldn't have any
dots in it. It should only be a LowerName and not a QualifiedLowerName.
Stuff like that.