Split up parser file
This might be overkill or maybe a bad choice but the parser file was
getting quite long and I saw another implementation that had the types
parsing split into a separate module but in order to do that it feels
better to separate out everything else too to limit circular imports.

Not in love with the 'mtch' name but 'match' is a keyword. Probably room
to improve but it is nice to have it a big broken out too.