Allow parsing of simple case-statement
This isn't able to evaluate the case-statement. Just parse it. The test
in tests/case.rs is currently failing as we've not reached the point of
having an acceptable snapshot.

I have once again revamped the indentation handling logic as I was quite
out of date with the old current/base pairing system and I haven't been
respecting it for a while. It also seems like it can definitely be
improved so we're at least not passing two values around everywhere.

Here we've refactored into an 'indentation' struct that attempts track
both the lines and spaces that it has encountered. The 'line' count is
not going to be perfect due to multi-line comments and mistakes but the
main idea is that is allows us to figure out if two different
indentation objects are referring to the same line or not.

Not all the code is coverted over yet. We'll get there hopefully.

The case-statement handling is very basic at the moment. It seems like
our use of the Pattern value for arguments in functions might need to be
tweaked. But we'll get there.