Allow checking & evaluating simple case statements
Very simple case statements. The checker isn't really tested to make
sure that it is validating anything at all.

We only support boolean case values at the moment. Plenty of room for
improvement.

We also needed to further improve the indent handling by always
returning the next_token_indent from the expression parsing as
previously it seems we were relying on the case-statement being at the
end of the file and so running out of tokens got us out of trouble but
with it moved into function we have to handle parsing the next statement
too.