Support underscore for matching in case-statements
We move the "pattern to term" logic out of the ast where it was a bit
odd and just have another converter function in the checker code.

We extend the lexer to support a single underscore as its own token
which means we can complain about it if found in certain code (ie. as an
arg to a function call). Hopefully it is the right thing to do.

I've also decided that a _ converts to an unique variable term which
might be an issue but as we can block it actually ever being used as a
value then it is probably fine.