Convert remaing whitespace processing
Over to using the base_indent approach instead of the 'base, current'
thingy.

Done just to clean up. I don't know that this is perfect though all the
tests pass for the moment. I'm sure there are plenty of issues lurking
to be discovered. It feels like we might need to pass some of the
indents back up the stack like we do with the expression parsing but I'm
not sure.